feat(processor): move unit-tests for line item mapping to separate file
